requirements.txt
click==6.7
Flask==0.12.2
Flask-Script==2.0.5
gunicorn==19.7.1
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
python-firebase==1.2
Werkzeug==0.12.2